UOE-9292: decouple pbs and sshb startup
