+++ main.go
@@ -1,30 +1,33 @@
-package prebidServer
+package main_ow
 
 import (
+	"flag"
 	"math/rand"
 	"net/http"
 	"path/filepath"
 	"runtime"
 	"time"
 
-	"github.com/golang/glog"
-	"github.com/prebid/prebid-server/usersync"
-
 	"github.com/prebid/prebid-server/config"
 	"github.com/prebid/prebid-server/currency"
 	"github.com/prebid/prebid-server/openrtb_ext"
 	"github.com/prebid/prebid-server/router"
+	"github.com/prebid/prebid-server/server"
 	"github.com/prebid/prebid-server/util/task"
 
+	"github.com/golang/glog"
 	"github.com/spf13/viper"
 )
 
-var InfoDirectory = "/home/http/GO_SERVER/dmhbserver/static/bidder-info"
-
-func InitPrebidServer(configFile string) {
+func init() {
 	rand.Seed(time.Now().UnixNano())
+}
 
-	bidderInfoPath, err := filepath.Abs(InfoDirectory)
+// TODO: revert this after PBS-OpenWrap module
+func Main() {
+	flag.Parse() // required for glog flags and testing package flags
+
+	bidderInfoPath, err := filepath.Abs(infoDirectory)
 	if err != nil {
 		glog.Exitf("Unable to build configuration directory path: %v", err)
 	}
@@ -33,7 +36,7 @@ func InitPrebidServer(configFile string) {
 	if err != nil {
 		glog.Exitf("Unable to load bidder configurations: %v", err)
 	}
-	cfg, err := loadConfig(bidderInfos, configFile)
+	cfg, err := loadConfig(bidderInfos)
 	if err != nil {
 		glog.Exitf("Configuration could not be loaded or did not pass validation: %v", err)
 	}
@@ -52,7 +55,10 @@ func InitPrebidServer(configFile string) {
 	}
 }
 
-func loadConfig(bidderInfos config.BidderInfos, configFileName string) (*config.Configuration, error) {
+const configFileName = "pbs.yaml"
+const infoDirectory = "./static/bidder-info"
+
+func loadConfig(bidderInfos config.BidderInfos) (*config.Configuration, error) {
 	v := viper.New()
 	config.SetupViper(v, configFileName, bidderInfos)
 	return config.New(v, bidderInfos, openrtb_ext.NormalizeBidderName)
@@ -66,34 +72,14 @@ func serve(cfg *config.Configuration) error {
 	currencyConverterTickerTask := task.NewTickerTask(fetchingInterval, currencyConverter)
 	currencyConverterTickerTask.Start()
 
-	_, err := router.New(cfg, currencyConverter)
+	r, err := router.New(cfg, currencyConverter)
 	if err != nil {
 		return err
 	}
 
+	corsRouter := router.SupportCORS(r)
+	server.Listen(cfg, router.NoCache{Handler: corsRouter}, router.Admin(currencyConverter, fetchingInterval), r.MetricsEngine)
+
+	r.Shutdown()
 	return nil
-}
-
-func OrtbAuction(w http.ResponseWriter, r *http.Request) error {
-	return router.OrtbAuctionEndpointWrapper(w, r)
-}
-
-var VideoAuction = func(w http.ResponseWriter, r *http.Request) error {
-	return router.VideoAuctionEndpointWrapper(w, r)
-}
-
-func GetUIDS(w http.ResponseWriter, r *http.Request) {
-	router.GetUIDSWrapper(w, r)
-}
-
-func SetUIDS(w http.ResponseWriter, r *http.Request) {
-	router.SetUIDSWrapper(w, r)
-}
-
-func CookieSync(w http.ResponseWriter, r *http.Request) {
-	router.CookieSync(w, r)
-}
-
-func SyncerMap() map[string]usersync.Syncer {
-	return router.SyncerMap()
 }
